chore: move bridge to features
