+++ apps/tangle-dapp/src/app/providers.tsx
@@ -7,9 +7,8 @@ import { UIProvider } from '@tangle-network/ui-components';
 import { type PropsWithChildren, type ReactNode } from 'react';
 import type { State } from 'wagmi';
 import { z } from 'zod';
-
-import HyperlaneWarpContext from '../context/bridge/HyperlaneWarpContext';
-import BridgeTxQueueProvider from '../context/bridge/BridgeTxQueueContext/BridgeTxQueueProvider';
+import HyperlaneWarpContext from '../features/bridge/context/HyperlaneWarpContext';
+import BridgeTxQueueProvider from '../features/bridge/context/BridgeTxQueueContext/BridgeTxQueueProvider';
 import PolkadotApiProvider from '@tangle-network/tangle-shared-ui/context/PolkadotApiProvider';
 
 const appEvent = new AppEvent();
