+++ apps/tangle-dapp/src/features/bridge/containers/BridgeContainer.tsx
@@ -23,30 +23,30 @@ import {
 import { useCallback, useEffect, useMemo, useState } from 'react';
 import { twMerge } from 'tailwind-merge';
 import { useShallow } from 'zustand/react/shallow';
-import AddressInput from '../../components/AddressInput';
-import AmountInput from '../../components/AmountInput';
-import { BridgeConfirmationModal } from '../../components/bridge/BridgeConfirmationModal';
-import { FeeDetail } from '../../components/bridge/FeeDetail';
-import ErrorMessage from '../../components/ErrorMessage';
-import { AssetConfig, AssetList } from '../../components/Lists/AssetList';
-import { ChainList } from '../../components/Lists/ChainList';
-import { AddressType } from '../../constants';
-import useBridgeStore from '../../context/bridge/useBridgeStore';
-import useBalances from '../../data/balances/useBalances';
-import { useBridgeEvmBalances } from '../../data/bridge/useBridgeEvmBalances';
-import { useHyperlaneQuote } from '../../data/bridge/useHyperlaneQuote';
-import useRouterQuote from '../../data/bridge/useRouterQuote';
-import useIsBridgeNativeToken from '../../hooks/useIsBridgeNativeToken';
+import AddressInput from '../../../components/AddressInput';
+import AmountInput from '../../../components/AmountInput';
+import { BridgeConfirmationModal } from '../components/BridgeConfirmationModal';
+import { FeeDetail } from '../components/FeeDetail';
+import ErrorMessage from '../../../components/ErrorMessage';
+import { AssetConfig, AssetList } from '../../../components/Lists/AssetList';
+import { ChainList } from '../../../components/Lists/ChainList';
+import { AddressType } from '../../../constants';
+import useBridgeStore from '../context/useBridgeStore';
+import useBalances from '../../../data/balances/useBalances';
+import { useBridgeEvmBalances } from '../hooks/useBridgeEvmBalances';
+import { useHyperlaneQuote } from '../hooks/useHyperlaneQuote';
+import useRouterQuote from '../hooks/useRouterQuote';
+import useIsBridgeNativeToken from '../hooks/useIsBridgeNativeToken';
 import { useWebContext } from '@tangle-network/api-provider-environment/webb-context';
 import useNetworkStore from '@tangle-network/tangle-shared-ui/context/useNetworkStore';
-import useBridgeAssets from '../../hooks/bridge/useBridgeAssets';
-import useBridgeTokenBalance from '../../hooks/bridge/useBridgeTokenBalance';
-import useHyperlaneFeeDetails from '../../hooks/bridge/useHyperlaneFeeDetails';
-import useRouterFeeDetails from '../../hooks/bridge/useRouterFeeDetails';
-import useRouterQuoteParams from '../../hooks/bridge/useRouterQuoteParams';
-import useHyperlaneQuoteParams from '../../hooks/bridge/useHyperlaneQuoteParams';
-import BridgeActionButton from '../../components/bridge/BridgeActionButton';
-import useRouterTransferData from '../../hooks/bridge/useRouterTransferData';
+import useBridgeAssets from '../hooks/useBridgeAssets';
+import useBridgeTokenBalance from '../hooks/useBridgeTokenBalance';
+import useHyperlaneFeeDetails from '../hooks/useHyperlaneFeeDetails';
+import useRouterFeeDetails from '../hooks/useRouterFeeDetails';
+import useRouterQuoteParams from '../hooks/useRouterQuoteParams';
+import useHyperlaneQuoteParams from '../hooks/useHyperlaneQuoteParams';
+import BridgeActionButton from '../components/BridgeActionButton';
+import useRouterTransferData from '../hooks/useRouterTransferData';
 import { get } from 'lodash';
 import { PresetTypedChainId } from '@tangle-network/dapp-types/ChainId';
