+++ apps/tangle-dapp/src/features/bridge/hooks/useHyperlaneTransfer.ts
@@ -2,13 +2,9 @@ import { TransactionReceipt } from '@ethersproject/abstract-provider';
 import { JsonRpcSigner } from '@ethersproject/providers';
 import { useMutation } from '@tanstack/react-query';
 import { providers, utils } from 'ethers';
-
 import { BridgeToken } from '@tangle-network/tangle-shared-ui/types';
-import { getHyperlaneWarpCore } from '../../lib/bridge/hyperlane/context';
-import {
-  getHyperlaneChainName,
-  tryFindToken,
-} from '../../lib/bridge/hyperlane/utils';
+import { getHyperlaneWarpCore } from '../lib/hyperlane/context';
+import { getHyperlaneChainName, tryFindToken } from '../lib/hyperlane/utils';
 import useEthersSigner from './useEthersSigner';
 import { PresetTypedChainId } from '@tangle-network/dapp-types';
