+++ apps/tangle-dapp/src/features/bridge/lib/hyperlane/assembleChainMetadata.ts
