+++ apps/tangle-dapp/src/features/bridge/lib/hyperlane/utils.ts
@@ -1,5 +1,4 @@
 import { IToken } from '@hyperlane-xyz/sdk';
-
 import { getHyperlaneWarpCore } from './context';
 import { PresetTypedChainId } from '@tangle-network/dapp-types';
